requirements.txt

discord.py==1.5.1
youtube_dl==2020.12.14
PyNaCl==1.4.0
loguru
